Moved login to platform.
