+++ src/js/disability-benefits/526EZ/components/VerifiedReviewContainer.jsx
@@ -23,7 +23,7 @@ import {
   saveAndRedirectToReturnUrl,
   autoSaveForm
 } from '../../../common/schemaform/save-in-progress/actions';
-import { toggleLoginModal } from '../../../login/actions';
+import { toggleLoginModal } from '../../../../platform/site-wide/login/actions';
 import {
   getNextPagePath,
   getPreviousPagePath
