+++ src/js/user-profile/reducers/profile.js
@@ -1,6 +1,6 @@
 import { merge, set } from 'lodash/fp';
 
-import { UPDATE_LOGGEDIN_STATUS } from '../../login/actions';
+import { UPDATE_LOGGEDIN_STATUS } from '../../../platform/site-wide/login/actions';
 
 import {
   UPDATE_PROFILE_FIELDS,
