+++ src/platform/site-wide/login/components/HelpMenu.jsx
@@ -1,7 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
-import DropDown from './DropDown';
+
 import IconHelp from '@department-of-veterans-affairs/jean-pants/IconHelp';
+import DropDown from '../../../../js/common/components/DropDown';
 
 class HelpMenu extends React.Component {
   render() {
