+++ src/platform/site-wide/login/login-entry.jsx
@@ -1,10 +1,8 @@
-import '../../sass/login.scss';
-
 import React from 'react';
-import startReactApp from '../../platform/startup/react';
-
 import { Provider } from 'react-redux';
 
+import '../../../sass/login.scss';
+import startReactApp from '../../startup/react';
 import Main from './containers/Main';
 
 export default function createLoginWidget(store) {
