+++ src/platform/site-wide/login/utils/helpers.js
@@ -1,9 +1,9 @@
 import Raven from 'raven-js';
 import appendQuery from 'append-query';
 
-import recordEvent from '../../../platform/monitoring/record-event';
-import { apiRequest } from '../../../platform/utilities/api';
-import environment from '../../../platform/utilities/environment';
+import recordEvent from '../../../monitoring/record-event';
+import { apiRequest } from '../../../utilities/api';
+import environment from '../../../utilities/environment';
 
 const SESSIONS_URI = `${environment.API_URL}/sessions`;
 const redirectUrl = (type) => `${SESSIONS_URI}/${type}/new`;
